% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{ModelResponse}
\alias{ModelResponse}
\title{Linear Model Response}
\description{
R6 object containing all necessary information of a linear model object
}
\section{Usage}{

\preformatted{
radon_model <- lm(Uppm ~ typebldg + basement + dupflag, data = radon_mn)
result <- extract_lm(radon_model)
result$as_json()
}
}

\section{Initialize}{

This should only be called by methods within the package
}

\section{Details}{

\code{$model_type}: short description of the model information. "linear model"

\code{$model_call}: character value of the call made to produce the model

\code{$response_variable}: character value of the response in the linear model

\code{$beta_variables}: character vector of beta variables used in the linear model

\code{$diag_model}: single row data.frame information about the linear model
Contains \describe{
    \item{\code{'r.squared'}}{The percent of variance explained by the model}
    \item{\code{'adj.r.squared'}}{r.squared adjusted based on the degrees of freedom}
    \item{\code{'sigma'}}{Estimate of residual standard deviation when corresponding observation is dropped from model}
    \item{\code{'statistic'}}{F-statistic}
    \item{\code{'p.value'}}{p-value from the F test, describing whether the full regression is significant}
    \item{\code{'df'}}{Degrees of freedom used by the coefficients}
    \item{\code{'logLik'}}{the data's log-likelihood under the model}
    \item{\code{'AIC'}}{the Akaike Information Criterion}
    \item{\code{'BIC'}}{the Bayesian Information Criterion}
    \item{\code{'deviance'}}{deviance}
    \item{\code{'df.residual'}}{residual degrees of freedom}
  }

\code{$diag_coefs}: 'p' row data.frame information about the linear model coefficients.
\describe{
  \item{\code{'term'}}{The term in the linear model being estimated and tested}
  \item{\code{'estimate'}}{The estimated coefficient}
  \item{\code{'std.error'}}{The standard error from the linear model}
  \item{\code{'statistic'}}{t-statistic}
  \item{\code{'p.value'}}{two-sided p-value}
}

\code{$diag_data}: 'n' row data.frame information about the linear model data.
\describe{
  \item{\code{'.fitted'}}{Fitted values of model}
  \item{\code{'.se.fit'}}{Standard errors of fitted values}
  \item{\code{'.resid'}}{Residuals}
  \item{\code{'.hat'}}{Diagonal of the hat matrix}
  \item{\code{'.sigma'}}{Estimate of residual standard deviation when corresponding observation is dropped from model}
  \item{\code{'.cooksd'}}{Cooks distance, ‘cooks.distance’}
  \item{\code{'.std.resid'}}{Standardised residuals. (Some unusual "lm" objects, such as "rlm" from MASS, may omit ‘.cooksd’ and ‘.std.resid’. "gam" from mgcv omits ‘.sigma’)}
}
}

