% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{predict_model}
\alias{predict_model}
\title{Predict a model}
\usage{
predict_model(id, data, predictor_variables)
}
\arguments{
\item{id}{model id that should exist in \code{models_by_id$ids()}}

\item{data}{data used in prediction}

\item{predictor_variables}{name of all predictor variables to be used. (character vector)... idk how this would really work}
}
\value{
json output of fitted values
}
\description{
Predict a model
}
