% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{run_model}
\alias{run_model}
\alias{run_lm}
\alias{run_quadratic}
\alias{run_loess}
\title{Run a model and extract features}
\usage{
run_model(data, response, predictor_variables, model_fn, model_name, ...)

run_lm(data, response, predictor_variables)

run_quadratic(data, response, predictor_variables, quadratic_variables)

run_loess(data, response, predictor_variables)
}
\arguments{
\item{data}{data used. (data.frame)}

\item{response}{single response variable to be used. (character)}

\item{predictor_variables}{name of all predictor variables to be used. (character vector)}

\item{model_fn}{function to create the model object}

\item{model_name}{string of function name to create the model object}

\item{...}{items passed to extract_model}

\item{quadratic_variables}{name of quadratic variables. (character vector)}
}
\description{
Run a model and extract features
}
\examples{
run_lm(iris, "Sepal.Length", c("Sepal.Width", "Petal.Width"))
run_quadratic(iris, "Sepal.Length", c("Sepal.Width", "Petal.Width"), c("Sepal.Width"))
run_loess(iris, "Sepal.Length", c("Sepal.Width", "Petal.Width"))
}
