% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{update_response}
\alias{update_response}
\alias{update_response.default}
\alias{update_response.lm}
\alias{update_response.loess}
\title{Update model response object}
\usage{
update_response(x, res, data, ...)

\method{update_response}{default}(x, res, data, ...)

\method{update_response}{lm}(x, res, data, ...)

\method{update_response}{loess}(x, res, data, ...)
}
\arguments{
\item{x}{model object}

\item{res}{response object of class \code{ModelResponse}}

\item{...}{ignored}
}
\description{
Update model response object
}
